requirements.txt
pydantic==2.4.2
py-solc-x==1.1.1
web3==6.0.0
